requirements.txt
django
djangorestframework
grpcio
google
protobuf